% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_gofLMM.R
\name{get.sim.proc.F}
\alias{get.sim.proc.F}
\title{Internal function}
\usage{
get.sim.proc.F(fit, residuals, std.type, use.correction.for.imbalance,
  subset.fix, order.by.original, or.original.fitted.P,
  or.original.fitted.S, original.fitted.P, original.fitted.S, n, N, x, ZZ,
  id, transform)
}
\description{
Internal function
}
\keyword{internal}
