% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_gofLMM.R
\name{get.sim.proc.F.sigma}
\alias{get.sim.proc.F.sigma}
\title{Internal function}
\usage{
get.sim.proc.F.sigma(
  fit,
  residuals,
  std.type,
  use.correction.for.imbalance,
  subset.fix,
  order.by.original,
  or.original.fitted.P,
  or.original.fitted.S,
  original.fitted.P,
  original.fitted.S,
  n,
  N,
  x,
  ZZ,
  id,
  sigmaP,
  sigmaPS
)
}
\description{
Internal function
}
\keyword{internal}
