% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_gofLMM.R
\name{get.sim.proc.fast.ororg}
\alias{get.sim.proc.fast.ororg}
\title{Internal function}
\usage{
get.sim.proc.fast.ororg(fit, std.type, use.correction.for.imbalance, n, N,
  x, ZZ, id, fittedI, or.fittedI, fittedP, or.fittedP)
}
\description{
Internal function
}
\keyword{internal}
