% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_gofLMM.R
\name{get.sim.proc.fast}
\alias{get.sim.proc.fast}
\title{Internal function}
\usage{
get.sim.proc.fast(fit, std.type, use.correction.for.imbalance, n, N, x, ZZ,
  id)
}
\description{
Internal function
}
\keyword{internal}
