% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_gofLMM.R
\name{get.sim.proc.O.sigma}
\alias{get.sim.proc.O.sigma}
\title{Internal function}
\usage{
get.sim.proc.O.sigma(fit, residuals, std.type,
  use.correction.for.imbalance, order.by.original, or.original.fitted.I,
  original.fitted.I, n, N, x, ZZ, id, sigmaI, transform)
}
\description{
Internal function
}
\keyword{internal}
