% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_gofLMM.R
\name{get.sim.proc.O.test.2}
\alias{get.sim.proc.O.test.2}
\title{Internal function}
\usage{
get.sim.proc.O.test.2(
  fit,
  residuals,
  std.type,
  use.correction.for.imbalance,
  order.by.original,
  n,
  N,
  x,
  ZZ,
  id,
  orest,
  ororder
)
}
\description{
Internal function
}
\keyword{internal}
