% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_gofLMM.R
\name{get.sim.proc.O.test}
\alias{get.sim.proc.O.test}
\title{Internal function}
\usage{
get.sim.proc.O.test(fit, residuals, std.type, use.correction.for.imbalance,
  order.by.original, n, N, x, ZZ, id, est.original, or.original)
}
\description{
Internal function
}
\keyword{internal}
