% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_gofLMM.R
\name{get.sim.proc.O.test.type2}
\alias{get.sim.proc.O.test.type2}
\title{Internal function}
\usage{
get.sim.proc.O.test.type2(fit, residuals, std.type,
  use.correction.for.imbalance, n, N, x, ZZ, id, est.original, or.original,
  A, B)
}
\description{
Internal function
}
\keyword{internal}
