% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_gofLMM.R
\name{gof.lmm.O.test.2}
\alias{gof.lmm.O.test.2}
\title{Goodness-of fit test for LMM}
\usage{
gof.lmm.O.test.2(
  fit,
  residuals = "individual",
  std.type = c(1, 2),
  use.correction.for.imbalance = FALSE,
  type = c("sign.flip", "permutation"),
  M = 100,
  order.by.original = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{fit}{The result of a call to \code{"nlme"}. The model must be fitted with \code{control=lmeControl( returnObject = TRUE)} and \code{keep.data=TRUE}. An error message is returned otherwise. ID variable must be numeric and ordered from 1:N ! Canno't use transofrmations of the outcome variable directly in the formula i.e. lme(sqrt(y)~x) will return p=1!}

\item{residuals}{Residuals to be used when constructing the process. Currently implemented only for \code{"individual"} for \textit{individual} residuals.}

\item{std.type}{Type of standardization to be used for the residuals when constructing the process.
Currently implemeneted options are \code{1} and \code{2} for $S_i=\hat\sigma^{-1/2}I_{n_i}$ and $S_i=\hat{V}_i^{-1/2}$.}

\item{use.correction.for.imbalance}{Logical. use $n_i^{-1/2} S_i$ when standardizing the residuals. Defaults to \code{FALSE}.}

\item{type}{How to obtain the processes $W^m$. Possible values are  \code{"sign.flip"} for the sign-flipping approach and \code{"permutation"} for the permutation approach.}

\item{M}{Number of random simulations/sign-flipps/permutations. Defaults to \code{100}.}

\item{order.by.original}{Order the residuals by the original fitted values. Deafults to FALSE.}

\item{verbose}{Logical. Print the current status of the test. Can slow down the algorithm, but it can make it feel faster. Defaults to \code{FALSE}.}
}
\description{
Goodness-of fit test based on cumulative sum stochastic process for O using a limit expressions for diagonal blocked matrices A and B.
}
\seealso{
\code{\link{gof.lmm}}
}
\author{
Rok Blagus, \email{rok.blagus@mf.uni-lj.si}
}
