% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_gofLMM.R
\name{gof.lmm.O.type2.boot}
\alias{gof.lmm.O.type2.boot}
\title{Goodness-of fit test for LMM}
\usage{
gof.lmm.O.type2.boot(fit, residuals = c("individual", "cluster"),
  std.type = c(1, 2), use.correction.for.imbalance = FALSE,
  type = c("sign.flip", "permutation"), M = 100, B = 100,
  order.by.original = FALSE, verbose = FALSE)
}
\arguments{
\item{fit}{The result of a call to \code{"nlme"}. The model must be fitted with \code{control=lmeControl( returnObject = TRUE)} and \code{keep.data=TRUE}. An error message is returned otherwise. ID variable must be numeric and ordered from 1:N ! Canno't use transofrmations of the outcome variable directly in the formula i.e. lme(sqrt(y)~x) will return p=1!}

\item{residuals}{Residuals to be used when constructing the process. Possible values are \code{"individual"} and \code{"cluster"} for \emph{individual} and \emph{cluster-speciffic} residuals, respectively.}

\item{std.type}{Type of standardization to be used for the residuals when constructing the process.
Currently implemeneted options are \code{1} and \code{2} for \eqn{S_i=\hat\sigma^{-1/2}I_{n_i}} and \eqn{S_i=\hat{V}_i^{-1/2}}.}

\item{use.correction.for.imbalance}{Logical. use \eqn{n_i^{-1/2} S_i} when standardizing the residuals. Defaults to \code{FALSE}.}

\item{type}{How to obtain the processes \eqn{W^m}. Possible values are  \code{"sign.flip"} for the sign-flipping approach and \code{"permutation"} for the permutation approach.}

\item{M}{Number of random simulations/sign-flipps/permutations. Defaults to \code{100}.}

\item{B}{Number of boot replications. Defaults to \code{100}.}

\item{order.by.original}{Logical. Should the residuals in the the processes \eqn{W^m} be ordered by the original fitted values? Defaults to \code{FALSE}.}

\item{verbose}{Logical. Print the current status of the test. Can slow down the algorithm, but it can make it feel faster. Defaults to \code{FALSE}.}
}
\value{
An object of class \code{"gofLMM"} for which \code{plot} and \code{summary} functions are available.
}
\description{
Goodness-of fit test based on cumulative sum stochastic process for Oi and John's idea using full parametric bootstrap to obtain p-values. Not well tested!
}
\seealso{
\code{\link{gof.lmm.pan}}, \code{\link{gof.lmm.sim}}
}
\author{
Rok Blagus, \email{rok.blagus@mf.uni-lj.si}
}
