% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_gofLMM.R
\name{gof.lmm.O.type2.i}
\alias{gof.lmm.O.type2.i}
\title{Internal function}
\usage{
gof.lmm.O.type2.i(
  fit,
  residuals,
  std.type,
  use.correction.for.imbalance,
  type,
  M,
  order.by.original,
  id,
  N,
  n,
  ZZ,
  x
)
}
\description{
Internal function
}
\keyword{internal}
