% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_gofLMM.R
\name{my.plot.gofLMM.part}
\alias{my.plot.gofLMM.part}
\title{Internal function}
\usage{
my.plot.gofLMM.part(W, Wm, type = c(1, 2), y, ym, ...)
}
\arguments{
\item{...}{additional arguments passed to from or to other methods}
}
\description{
Internal function
}
\keyword{internal}
