% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_gofLMM.R
\name{my.summary.gofLMM.testO}
\alias{my.summary.gofLMM.testO}
\title{Summary function}
\usage{
my.summary.gofLMM.testO(object, ...)
}
\arguments{
\item{object}{an object returned by a call to a function}

\item{...}{additional arguments passed to from or to other methods}
}
\description{
summary function
}
