% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_gofLMM.R
\name{plot.gofLMM}
\alias{plot.gofLMM}
\title{Plot Function}
\usage{
\method{plot}{gofLMM}(object, type = 2, subset.M = NULL, ...)
}
\arguments{
\item{object}{an object of class \code{"gofLMM"}, an object returned by a call to \code{\link{gof.lmm}}}

\item{type}{Type of x-axis. Possible values are 1 for 1:N and 2 for the predicted values. Defaults to 2.}

\item{subset.M}{How many realizations of $W^m$ should be plotted. Defaults to NULL and all the realizations are plotted.}

\item{...}{additional arguments passed to from or to other methods}
}
\description{
plots the processes which are the result of a call to \code{gof.lmm}
}
\seealso{
\code{\link{gof.lmm.pan}}, \code{\link{gof.lmm}} and \code{\link{summary.gofLMM}}
}
\author{
Rok Blagus, \email{rok.blagus@mf.uni-lj.si}
}
