% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_gofLMM.R
\name{print.gofLMM}
\alias{print.gofLMM}
\title{Print Function}
\usage{
\method{print}{gofLMM}(x, ...)
}
\arguments{
\item{x}{an object of class \code{"gofLMM"}, an object returned by a call to \code{\link{gof.lmm}}}

\item{...}{additional arguments passed to from or to other methods}
}
\description{
prints results from a call to \code{gof.lmm}
}
\seealso{
\code{\link{gof.lmm}}, \code{\link{summary.gofLMM}} and \code{\link{plot.gofLMM}}
}
\author{
Rok Blagus, \email{rok.blagus@mf.uni-lj.si}
}
