% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_gofLMM.R
\name{summary.gofLMM}
\alias{summary.gofLMM}
\title{Summary Function}
\usage{
\method{summary}{gofLMM}(object, ...)
}
\arguments{
\item{object}{an object of class \code{"gofLMM"}, an object returned by a call to \code{\link{gof.lmm}}}

\item{...}{additional arguments passed to from or to other methods}
}
\value{
a matrix containing KS and CvM test statistics and corresponding \eqn{p}-values for the constructed processes.
}
\description{
makes a summary of a call to \code{gof.lmm}
}
\seealso{
\code{\link{gof.lmm.pan}}, \code{\link{gof.lmm}} and \code{\link{plot.gofLMM}}
}
\author{
Rok Blagus, \email{rok.blagus@mf.uni-lj.si}
}
